# api-backend/fripay/public/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FriPay API Documentation",
        "version": "1.0.0",
        "description": "API FriPay - Transferts inter-op\u00e9rateurs Mobile Money pour le B\u00e9nin. Endpoint de documentation: `\/api\/docs\/api`"
    },
    "servers": [
        {
            "url": "http:\/\/payments.fripay.local\/api"
        }
    ],
    "paths": {
...
        "\/v1\/webhooks\/aggregator\/{provider}": {
            "post": {
                "operationId": "webhook.handleAggregator",
                "summary": "POST \/api\/v1\/webhooks\/aggregator\/{provider}",
                "tags": [
                    "Webhook"
                ],
                "parameters": [
                    {
                        "name": "provider",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "const": "ok"
                                        }
                                    },
                                    "required": [
                                        "status"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "type": {
                                            "type": "string",
                                            "const": "INVALID_SIGNATURE"
                                        },
                                        "title": {
                                            "type": "string",
                                            "const": "Signature invalide"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "const": 401
                                        },
                                        "detail": {
                                            "type": "string",
                                            "const": "La signature du webhook est invalide."
                                        },
                                        "request_id": {
                                            "anyOf": [
                                                {
                                                    "type": "string"
                                                },
                                                {
                                                    "type": "array",
                                                    "items": {
                                                        "__PHP_Incomplete_Class_Name": "stdClass"
                                                    }
                                                },
                                                {
                                                    "type": "null"
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "type",
                                        "title",
                                        "status",
                                        "detail",
                                        "request_id"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "\/v1\/webhooks\/pispi": {
            "post": {
                "operationId": "webhook.handlePispi",
                "summary": "POST \/api\/v1\/webhooks\/pispi",
                "tags": [
                    "Webhook"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "const": "ok"
                                        }
                                    },
                                    "required": [
                                        "status"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "type": {
                                            "type": "string",
                                            "const": "INVALID_SIGNATURE"
                                        },
                                        "title": {
                                            "type": "string",
                                            "const": "Signature invalide"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "const": 401
                                        },
                                        "detail": {
                                            "type": "string",
                                            "const": "La signature du webhook est invalide."
                                        },
                                        "request_id": {
                                            "anyOf": [
                                                {
                                                    "type": "string"
                                                },
                                                {
                                                    "type": "array",
                                                    "items": {
                                                        "__PHP_Incomplete_Class_Name": "stdClass"
                                                    }
                                                },
                                                {
                                                    "type": "null"
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "type",
                                        "title",
                                        "status",
                                        "detail",
                                        "request_id"
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}